+++ src/Sum.java
@@ -6,13 +6,13 @@ public class Sum extends Function {
     public String derivative() {
 
     }
-    public double bisectionMethod(int a, int b, int epsilon) {
+    public double bisectionMethod(int a, int b, double epsilon) {
 
     }
     public double bisectionMethod(int a, int b) {
 
     }
-    public double newtonRaphsonMethod(int a, int epsilon) {
+    public double newtonRaphsonMethod(int a, double epsilon) {
 
     }
     public double newtonRaphsonMethod(int a) {
